list_gemini_models.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")

if GOOGLE_API_KEY:
    genai.configure(api_key=GOOGLE_API_KEY)
    for m in genai.list_models():
        print(m.name)
else:
    print("GOOGLE_API_KEY not found in environment variables.")
