requirements.txt
pandas
numpy
serpapi
langgraph
yfinance
openai
python-dotenv
google-generativeai
matplotlib
